HTML-788: Fix failing tests and refactor from 2.3 -> 2.4
FormRecordable is only available in 2.4
