+++ api-2.4/pom.xml
@@ -7,14 +7,14 @@
         <version>4.4.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>htmlformentry-api-2.3</artifactId>
+    <artifactId>htmlformentry-api-2.4</artifactId>
     <packaging>jar</packaging>
-    <name>HTML Form Entry API 2.3</name>
-    <description>2.3 api project for htmlformentry</description>
+    <name>HTML Form Entry API 2.4</name>
+    <description>2.4 api project for htmlformentry</description>
 
     <properties>
-        <openMRSVersion>2.3.3</openMRSVersion>
-		<openMRSMinorVersion>2.3</openMRSMinorVersion>
+        <openMRSVersion>2.4.0</openMRSVersion>
+		<openMRSMinorVersion>2.4</openMRSMinorVersion>
     </properties>
 
     <dependencies>
