+++ api-2.4/src/main/java/org/openmrs/module/htmlformentry/FormEntryContextFactoryImpl2_4.java
@@ -4,11 +4,11 @@ import org.openmrs.annotation.OpenmrsProfile;
 import org.springframework.stereotype.Component;
 
 @Component("formEntryContextFactoryImpl")
-@OpenmrsProfile(openmrsPlatformVersion = "2.3.* - 2.*")
-public class FormEntryContextFactoryImpl2_3 implements FormEntryContextFactory {
+@OpenmrsProfile(openmrsPlatformVersion = "2.4.* - 2.*")
+public class FormEntryContextFactoryImpl2_4 implements FormEntryContextFactory {
 	
 	@Override
 	public FormEntryContext create(FormEntryContext.Mode mode) {
-		return new FormEntryContext2_3(mode);
+		return new FormEntryContext2_4(mode);
 	}
 }
