+++ api-2.4/src/test/java/org/openmrs/htmlformentry/HtmlFormEntryUtil2_4Test.java
@@ -6,11 +6,11 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.openmrs.Obs;
-import org.openmrs.module.htmlformentry.HtmlFormEntryUtil2_3;
+import org.openmrs.module.htmlformentry.HtmlFormEntryUtil2_4;
 import org.powermock.modules.junit4.PowerMockRunner;
 
 @RunWith(PowerMockRunner.class)
-public class HtmlFormEntryUtil2_3Test {
+public class HtmlFormEntryUtil2_4Test {
 	
 	@Test
 	public void getControlId_shouldReturnControlId() {
@@ -20,7 +20,7 @@ public class HtmlFormEntryUtil2_3Test {
 		observation.setFormField(FORM_NAMESPACE, "MyForm.1.0/my_condition_tag-0");
 		
 		// Test
-		String controlId = HtmlFormEntryUtil2_3.getControlId(observation);
+		String controlId = HtmlFormEntryUtil2_4.getControlId(observation);
 		
 		// Validation
 		Assert.assertEquals("my_condition_tag", controlId);
@@ -34,7 +34,7 @@ public class HtmlFormEntryUtil2_3Test {
 		observation.setFormField(FORM_NAMESPACE, "MyForm.1.0/my-condition-tag-0");
 		
 		// Test
-		String controlId = HtmlFormEntryUtil2_3.getControlId(observation);
+		String controlId = HtmlFormEntryUtil2_4.getControlId(observation);
 		
 		// Validation
 		Assert.assertEquals("my-condition-tag", controlId);
@@ -48,7 +48,7 @@ public class HtmlFormEntryUtil2_3Test {
 		observation.setFormField(FORM_NAMESPACE, "MyForm.1.0/my_condition_tag-X");
 		
 		// Test
-		String controlId = HtmlFormEntryUtil2_3.getControlId(observation);
+		String controlId = HtmlFormEntryUtil2_4.getControlId(observation);
 		
 		// Validation
 		Assert.assertEquals("my_condition_tag-X", controlId);
@@ -62,7 +62,7 @@ public class HtmlFormEntryUtil2_3Test {
 		observation.setFormField(FORM_NAMESPACE, "MyForm.1.0/my_condition_tag-123-0");
 		
 		// Test
-		String controlId = HtmlFormEntryUtil2_3.getControlId(observation);
+		String controlId = HtmlFormEntryUtil2_4.getControlId(observation);
 		
 		// Validation
 		Assert.assertEquals("my_condition_tag-123", controlId);
@@ -76,7 +76,7 @@ public class HtmlFormEntryUtil2_3Test {
 		observation.setFormField(FORM_NAMESPACE, "MyForm.1.0-42fcd95f/my_condition_tag-0");
 		
 		// Test
-		String controlId = HtmlFormEntryUtil2_3.getControlId(observation);
+		String controlId = HtmlFormEntryUtil2_4.getControlId(observation);
 		
 		// Validation
 		Assert.assertEquals("my_condition_tag", controlId);
@@ -89,7 +89,7 @@ public class HtmlFormEntryUtil2_3Test {
 		Obs observation = new Obs();
 		
 		// Test
-		String controlId = HtmlFormEntryUtil2_3.getControlId(observation);
+		String controlId = HtmlFormEntryUtil2_4.getControlId(observation);
 		
 		// Validation
 		Assert.assertNull(controlId);
