+++ api-2.4/src/test/resources/TestingApplicationContext.xml
