+++ api-2.4/src/test/resources/test-hibernate.cfg.xml
