+++ api-tests/pom.xml
@@ -149,4 +149,29 @@
 		<MODULE_PACKAGE>${project.parent.groupId}.${project.parent.artifactId}</MODULE_PACKAGE>
 	</properties>
 
+	<profiles>
+		<profile>
+			<id>2.4</id>
+			<dependencies>
+				<dependency>
+					<groupId>${project.groupId}</groupId>
+					<artifactId>${project.parent.artifactId}-api-2.4</artifactId>
+					<version>${project.parent.version}</version>
+					<scope>test</scope>
+				</dependency>
+			</dependencies>
+		</profile>
+		<profile>
+			<id>2.5</id>
+			<dependencies>
+				<dependency>
+					<groupId>${project.groupId}</groupId>
+					<artifactId>${project.parent.artifactId}-api-2.4</artifactId>
+					<version>${project.parent.version}</version>
+					<scope>test</scope>
+				</dependency>
+			</dependencies>
+		</profile>
+	</profiles>
+
 </project>
