+++ api-tests/src/test/java/org/openmrs/module/htmlformentry/MetadataSharingExportTest.java
@@ -28,7 +28,7 @@ public class MetadataSharingExportTest extends BaseHtmlFormEntryTest {
 	
 	@Before
 	public void setup() throws Exception {
-		executeVersionedDataSet("org/openmrs/module/htmlformentry/data/encounterProviderAndRole.xml");
+		executeVersionedDataSet("org/openmres/module/htmlformentry/data/encounterProviderAndRole.xml");
 		new HtmlFormEntryActivator().started();
 	}
