+++ api/src/main/java/org/openmrs/module/htmlformentry/FormEntryContextFactoryImpl.java
@@ -4,7 +4,7 @@ import org.openmrs.annotation.OpenmrsProfile;
 import org.springframework.stereotype.Component;
 
 @Component
-@OpenmrsProfile(openmrsPlatformVersion = "1.* - 2.2.*")
+@OpenmrsProfile(openmrsPlatformVersion = "1.* - 2.3.*")
 public class FormEntryContextFactoryImpl implements FormEntryContextFactory {
 	
 	@Override
